Aux window: validate window bounds are on display bounds (fix #204346) (#204399)
