+++ src/vs/platform/windows/electron-main/windows.ts
@@ -3,13 +3,13 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { BrowserWindowConstructorOptions, WebContents } from 'electron';
+import { BrowserWindowConstructorOptions, Display, Rectangle, WebContents, screen } from 'electron';
 import { Event } from 'vs/base/common/event';
 import { IProcessEnvironment, isLinux, isMacintosh, isWindows } from 'vs/base/common/platform';
 import { URI } from 'vs/base/common/uri';
 import { NativeParsedArgs } from 'vs/platform/environment/common/argv';
 import { ServicesAccessor, createDecorator } from 'vs/platform/instantiation/common/instantiation';
-import { ICodeWindow, IWindowState } from 'vs/platform/window/electron-main/window';
+import { ICodeWindow, IWindowState, WindowMode, defaultWindowState } from 'vs/platform/window/electron-main/window';
 import { IOpenEmptyWindowOptions, IWindowOpenable, IWindowSettings, WindowMinimumSize, hasNativeTitlebar, useNativeFullScreen, useWindowControlsOverlay, zoomLevelToZoomFactor } from 'vs/platform/window/common/window';
 import { IThemeMainService } from 'vs/platform/theme/electron-main/themeMainService';
 import { IProductService } from 'vs/platform/product/common/productService';
@@ -18,6 +18,7 @@ import { IEnvironmentMainService } from 'vs/platform/environment/electron-main/e
 import { join } from 'vs/base/common/path';
 import { IAuxiliaryWindow } from 'vs/platform/auxiliaryWindow/electron-main/auxiliaryWindow';
 import { Color } from 'vs/base/common/color';
+import { ILogService } from 'vs/platform/log/common/log';
 
 export const IWindowsMainService = createDecorator<IWindowsMainService>('windowsMainService');
 
@@ -213,4 +214,149 @@ export function getLastFocused(windows: ICodeWindow[] | IAuxiliaryWindow[]): ICo
 	}
 
 	return lastFocusedWindow;
+}
+
+export namespace WindowStateValidator {
+
+	export function validateWindowState(logService: ILogService, state: IWindowState, displays = screen.getAllDisplays()): IWindowState | undefined {
+		logService.trace(`window#validateWindowState: validating window state on ${displays.length} display(s)`, state);
+
+		if (
+			typeof state.x !== 'number' ||
+			typeof state.y !== 'number' ||
+			typeof state.width !== 'number' ||
+			typeof state.height !== 'number'
+		) {
+			logService.trace('window#validateWindowState: unexpected type of state values');
+
+			return undefined;
+		}
+
+		if (state.width <= 0 || state.height <= 0) {
+			logService.trace('window#validateWindowState: unexpected negative values');
+
+			return undefined;
+		}
+
+		// Single Monitor: be strict about x/y positioning
+		// macOS & Linux: these OS seem to be pretty good in ensuring that a window is never outside of it's bounds.
+		// Windows: it is possible to have a window with a size that makes it fall out of the window. our strategy
+		//          is to try as much as possible to keep the window in the monitor bounds. we are not as strict as
+		//          macOS and Linux and allow the window to exceed the monitor bounds as long as the window is still
+		//          some pixels (128) visible on the screen for the user to drag it back.
+		if (displays.length === 1) {
+			const displayWorkingArea = getWorkingArea(displays[0]);
+			if (displayWorkingArea) {
+				logService.trace('window#validateWindowState: 1 monitor working area', displayWorkingArea);
+
+				function ensureStateInDisplayWorkingArea(): void {
+					if (!state || typeof state.x !== 'number' || typeof state.y !== 'number' || !displayWorkingArea) {
+						return;
+					}
+
+					if (state.x < displayWorkingArea.x) {
+						// prevent window from falling out of the screen to the left
+						state.x = displayWorkingArea.x;
+					}
+
+					if (state.y < displayWorkingArea.y) {
+						// prevent window from falling out of the screen to the top
+						state.y = displayWorkingArea.y;
+					}
+				}
+
+				// ensure state is not outside display working area (top, left)
+				ensureStateInDisplayWorkingArea();
+
+				if (state.width > displayWorkingArea.width) {
+					// prevent window from exceeding display bounds width
+					state.width = displayWorkingArea.width;
+				}
+
+				if (state.height > displayWorkingArea.height) {
+					// prevent window from exceeding display bounds height
+					state.height = displayWorkingArea.height;
+				}
+
+				if (state.x > (displayWorkingArea.x + displayWorkingArea.width - 128)) {
+					// prevent window from falling out of the screen to the right with
+					// 128px margin by positioning the window to the far right edge of
+					// the screen
+					state.x = displayWorkingArea.x + displayWorkingArea.width - state.width;
+				}
+
+				if (state.y > (displayWorkingArea.y + displayWorkingArea.height - 128)) {
+					// prevent window from falling out of the screen to the bottom with
+					// 128px margin by positioning the window to the far bottom edge of
+					// the screen
+					state.y = displayWorkingArea.y + displayWorkingArea.height - state.height;
+				}
+
+				// again ensure state is not outside display working area
+				// (it may have changed from the previous validation step)
+				ensureStateInDisplayWorkingArea();
+			}
+
+			return state;
+		}
+
+		// Multi Montior (fullscreen): try to find the previously used display
+		if (state.display && state.mode === WindowMode.Fullscreen) {
+			const display = displays.find(d => d.id === state.display);
+			if (display && typeof display.bounds?.x === 'number' && typeof display.bounds?.y === 'number') {
+				logService.trace('window#validateWindowState: restoring fullscreen to previous display');
+
+				const defaults = defaultWindowState(WindowMode.Fullscreen); // make sure we have good values when the user restores the window
+				defaults.x = display.bounds.x; // carefull to use displays x/y position so that the window ends up on the correct monitor
+				defaults.y = display.bounds.y;
+
+				return defaults;
+			}
+		}
+
+		// Multi Monitor (non-fullscreen): ensure window is within display bounds
+		let display: Display | undefined;
+		let displayWorkingArea: Rectangle | undefined;
+		try {
+			display = screen.getDisplayMatching({ x: state.x, y: state.y, width: state.width, height: state.height });
+			displayWorkingArea = getWorkingArea(display);
+		} catch (error) {
+			// Electron has weird conditions under which it throws errors
+			// e.g. https://github.com/microsoft/vscode/issues/100334 when
+			// large numbers are passed in
+		}
+
+		if (
+			display &&														// we have a display matching the desired bounds
+			displayWorkingArea &&											// we have valid working area bounds
+			state.x + state.width > displayWorkingArea.x &&					// prevent window from falling out of the screen to the left
+			state.y + state.height > displayWorkingArea.y &&				// prevent window from falling out of the screen to the top
+			state.x < displayWorkingArea.x + displayWorkingArea.width &&	// prevent window from falling out of the screen to the right
+			state.y < displayWorkingArea.y + displayWorkingArea.height		// prevent window from falling out of the screen to the bottom
+		) {
+			logService.trace('window#validateWindowState: multi-monitor working area', displayWorkingArea);
+
+			return state;
+		}
+
+		return undefined;
+	}
+
+	function getWorkingArea(display: Display): Rectangle | undefined {
+
+		// Prefer the working area of the display to account for taskbars on the
+		// desktop being positioned somewhere (https://github.com/microsoft/vscode/issues/50830).
+		//
+		// Linux X11 sessions sometimes report wrong display bounds, so we validate
+		// the reported sizes are positive.
+		if (display.workArea.width > 0 && display.workArea.height > 0) {
+			return display.workArea;
+		}
+
+		if (display.bounds.width > 0 && display.bounds.height > 0) {
+			return display.bounds;
+		}
+
+		return undefined;
+	}
 }
