+++ src/vs/workbench/contrib/url/browser/externalUriResolver.ts
@@ -18,13 +18,13 @@ export class ExternalUriResolverContribution extends Disposable implements IWork
 	) {
 		super();
 
-		if (_workbenchEnvironmentService.options && _workbenchEnvironmentService.options.resolveExternalUri) {
+		if (_workbenchEnvironmentService.options?.resolveExternalUri) {
 			this._register(_openerService.registerExternalUriResolver({
 				resolveExternalUri: async (resource) => {
 					return {
 						resolved: await _workbenchEnvironmentService.options!.resolveExternalUri!(resource),
 						dispose: () => {
-							// TODO
+							// TODO@mjbvz - do we need to do anything here?
 						}
 					};
 				}
